refactor/fetch 로직 커스텀 훅으로 리펙토링
