+++ src/pages/ProductDetail.tsx
@@ -1,5 +1,4 @@
 import { useParams } from "react-router-dom";
-import { useEffect, useState } from "react";
 
 import { useRecoilValue } from "recoil";
 import { categoriesAtom } from "../atoms";
@@ -8,66 +7,16 @@ import ProductDetailSkeletonCard from "../components/skeleton/ProductDetailSkele
 
 import ProductDetailContainer from "../components/ProductDetail/ProductDetailContainer";
 
+import useProductFetch from "../hooks/useProductFetch";
+
 type ProductDetailProps = {
   children?: React.ReactNode;
 };
 
-interface ProductInfo {
-  id: number;
-  title: string;
-  price: number;
-  description: string;
-  category: string;
-  image: string;
-  rating: { rate: number; count: number };
-}
-
-const API_URL = "https://fakestoreapi.com/products";
-
-const STAR_NUM = 5;
-
 export default function ProductDetail(props: ProductDetailProps) {
-  const [data, setData] = useState<ProductInfo>({
-    id: 0,
-    title: "",
-    price: 0,
-    description: "",
-    category: "",
-    image: "",
-    rating: { rate: 0, count: 0 },
-  });
-
   const categories = useRecoilValue(categoriesAtom);
-  const [clicked, setClicked] = useState(new Array(STAR_NUM).fill(false));
-  const [loading, setLoading] = useState(false);
-
   let productId = useParams().productId;
-
-  const getData = () => {
-    fetch(`${API_URL}/${productId}`)
-      .then((res) => res.json())
-      .then((newData) => {
-        setData(newData);
-        return newData;
-      })
-      .then((data) => {
-        let num = Math.round(data.rating.rate);
-        let clickStates = new Array(STAR_NUM).fill(false);
-
-        for (let i = 0; i < num; i++) {
-          clickStates[i] = true;
-        }
-        // console.log(clickStates);
-        setClicked(() => clickStates);
-        setTimeout(() => setLoading(false), 1000);
-      })
-      .catch((err) => console.log(err));
-  };
-
-  useEffect(() => {
-    setLoading(true);
-    getData();
-  }, [productId]);
+  const { data, loading, clicked } = useProductFetch(productId as string);
 
   const findCategory = () => {
     let category = "";
